RegistrationAPI for NMOS v1.2
