+++ model/index.js
@@ -1,0 +1,11 @@
+
+var models = {
+  Node : require('./Node.js'),
+  Device : require('./Device.js'),
+  Source : require('./Source.js'),
+  Flow : require('./Flow.js'),
+  Sender : require('./Sender.js'),
+  Receiver : require('./Receiver.js'),
+};
+
+module.exports = models;
