+++ package.json
@@ -3,45 +3,25 @@
   "version": "0.0.1",
   "description": "Node.js Module for NMOS IS-04 Node, Discovery, Registration & Query",
   "main": "index.js",
-  "bin": {
-    "nmos-ledger": "./bin/nmos-ledger"
-  },
-  "directories": {
-    "test": "test"
-  },
   "dependencies": {
-    "ajv": "^6.10.2",
-    "async": "^2.6.0",
     "axios": "^0.19.0",
     "body-parser": "^1.18.2",
     "express": "^4.16.2",
     "lodash": "^4.17.15",
     "mdns": "^2.5.1",
-    "mdns-js": "git+https://github.com/Streampunk/nmos-mdns-js#master",
-    "promise": "^8.0.1",
-    "request": "^2.88.0",
-    "seamless-immutable": "^7.1.3",
-    "uuid": "^2.0.1",
-    "ws": "^5.1.1"
+    "uuid": "^2.0.1"
   },
   "devDependencies": {
     "tape": "^4.9.0"
   },
   "scripts": {
-    "test": "tape test/integration/*.js test/*.js"
+    "test": "tape test/integration/*.js test/*.js",
+    "start": "node app.js"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/jlommori/node-nmos-is04"
   },
-  "keywords": [
-    "AMWA",
-    "incubator",
-    "registration",
-    "discovery",
-    "JT-NM",
-    "NMOS"
-  ],
   "author": "Jeremy Lommori",
   "license": "Apache-2.0",
   "bugs": {
